requirements.txt
absl-py==2.1.0
accelerate==0.27.2
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anylabeling==0.2.23
argcomplete==3.2.2
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.22.1
blessed==1.20.0
blessings==1.7
blis==0.7.11
catalogue==2.0.10
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpathlib==0.16.0
cmake==3.28.3
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cryptography==40.0.2
cuda-python==12.3.0
cycler==0.12.1
cymem==2.0.8
datasets==2.17.1
datumaro==1.2.1
debugpy==1.8.1
decorator==5.1.1
diffusers==0.26.3
dill==0.3.8
einops==0.3.0
evaluate==0.4.1
exceptiongroup==1.1.1
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
future==0.18.3
gdown==5.1.0
geffnet==1.0.2
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imgviz==1.7.2
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.14.0
kiwisolver==1.4.5
kornia==0.6.0
langcodes==3.3.0
lazy_loader==0.3
lit==17.0.6
llvmlite==0.40.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.7.1
mmdet==2.28.0
mmengine==0.8.4
mmsegmentation==0.30.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
natsort==8.3.1
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.1.0
nltk==3.8.1
numba==0.57.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.1.1
onnx==1.13.1
onnxruntime==1.14.1
open-clip-torch==2.0.2
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
openvino==2022.3.0
optimum==1.17.1
orjson==3.8.11
outcome==1.2.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
path==16.10.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==3.10.0
pooch==1.7.0
preshed==3.0.9
prettytable==3.10.0
progressbar==2.5
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyarrow-hotfix==0.6
pycocotools==2.0.7
pydantic==2.6.4
pydantic_core==2.16.3
pyDeprecate==0.3.1
pyemd==1.0.0
Pygments==2.17.2
PyMatting==1.1.8
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
qimage2ndarray==1.10.0
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rich==13.7.0
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stable-diffusion-sdkit==2.1.5
stack-data==0.6.3
supervision==0.6.0
sympy==1.12
taming-transformers-rom1504==0.0.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6
terminaltables==3.1.10
test_tube==0.7.5
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchmetrics==0.6.0
torchvision==0.15.2
tornado==6.4
traitlets==5.14.1
transformers==4.33.2
trio==0.22.0
trio-websocket==0.10.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wsproto==1.2.0
xformers==0.0.22
xxhash==3.4.1
yapf==0.40.1
yarl==1.9.4
zipp==3.17.0
